feat(): AdminDataController - restFull

+++ src/main/java/com/infosystem/springmvc/controller/AdminDataController.java
@@ -4,35 +4,45 @@ import com.infosystem.springmvc.dto.*;
 import com.infosystem.springmvc.exception.DatabaseException;
 import com.infosystem.springmvc.exception.LogicException;
 import com.infosystem.springmvc.exception.ValidationException;
-import com.infosystem.springmvc.model.entity.User;
 import com.infosystem.springmvc.service.ContractService;
 import com.infosystem.springmvc.service.TariffOptionService;
 import com.infosystem.springmvc.service.TariffService;
+import com.infosystem.springmvc.validators.SearchUserByNumberValidator;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.MessageSource;
 import org.springframework.http.MediaType;
+import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.BindingResult;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
 
 import com.infosystem.springmvc.service.UserService;
 
 import javax.validation.Valid;
+import java.util.Locale;
 
 
 @RestController
-public class AdminDataController {
+public class AdminDataController extends ControllerTemplate {
 
     private final UserService userService;
     private final ContractService contractService;
     private final TariffOptionService tariffOptionService;
     private final TariffService tariffService;
+    private final SearchUserByNumberValidator searchUserByNumberValidator;
+    private final MessageSource messageSource;
 
     @Autowired
     public AdminDataController(UserService userService, ContractService contractService,
-                               TariffOptionService tariffOptionService, TariffService tariffService) {
+                               TariffOptionService tariffOptionService, TariffService tariffService,
+                               SearchUserByNumberValidator searchUserByNumberValidator,
+                               MessageSource messageSource) {
         this.userService = userService;
         this.contractService = contractService;
         this.tariffOptionService = tariffOptionService;
         this.tariffService = tariffService;
+        this.searchUserByNumberValidator = searchUserByNumberValidator;
+        this.messageSource = messageSource;
     }
 
     /**
@@ -46,12 +56,12 @@ public class AdminDataController {
      */
     @RequestMapping(value = {"/adminPanel/addContract", "/adminPanel/addContractToUser/{userId}"}, consumes = MediaType.APPLICATION_JSON_UTF8_VALUE,
             method = RequestMethod.POST)
-    public String addContract(@Valid @RequestBody AddContractDto addContractDto, BindingResult result) throws LogicException, ValidationException, DatabaseException {
+    public ResponseDto addContract(@Valid @RequestBody AddContractDto addContractDto, BindingResult result) throws LogicException, ValidationException, DatabaseException {
         if (result.hasErrors()) {
             throw new ValidationException("Enter correct phone number (length: 6-32) and chose tariff please.");
         }
         contractService.newContract(addContractDto);
-        return "Contract added successfully.";
+        return new ResponseDto("Contract added successfully.");
     }
 
     /**
@@ -64,33 +74,37 @@ public class AdminDataController {
      * @throws ValidationException if data in fields is not valid (tariff name length(min = 2, max = 32), min price = 1)
      */
     @RequestMapping(value = "/adminPanel/addTariff", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String saveTariff(@Valid @RequestBody AddTariffDto addTariffDto, BindingResult result) throws LogicException, ValidationException {
+    public ResponseDto saveTariff(@Valid @RequestBody AddTariffDto addTariffDto, BindingResult result) throws LogicException, ValidationException {
         if (result.hasErrors()) {
             throw new ValidationException("Enter correct name and price for new tariff please (name length between 2 and 32, min price 1).");
         }
         tariffService.addTariff(addTariffDto);
-        return "Tariff added successfully.";
+        return new ResponseDto("Tariff added successfully.");
     }
 
     /**
      * Called while searching user by number (adminPanel and allUsers pages).
      *
-     * @param searchByNumberDto searchByNumberDto
-     * @param result         validation result
+     * @param phoneNumber phoneNumber
      * @return message
      * @throws LogicException      if no such number
      * @throws ValidationException if data in fields is not valid (phone number length(min = 6, max = 32))
      */
-    @RequestMapping(value = "/adminPanel/user/searchUserByNumber", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE,
-            produces = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public @ResponseBody
-    Integer searchUserByNumber(@Valid @RequestBody SearchByNumberDto searchByNumberDto, BindingResult result) throws ValidationException, LogicException {
-        if (result.hasErrors()) {
-            throw new ValidationException("Phone number length is between 6 and 32 (only numbers).");
+    @RequestMapping(value = "/adminPanel/user/searchUserByNumber/{phoneNumber}", produces = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public SearchByNumberResponseDto searchUserByNumber(@PathVariable String phoneNumber) throws ValidationException, LogicException {
+        if (!pathVariableIsANumber(phoneNumber)) {
+            throw new ValidationException("Wrong path variable.");
         }
-        return userService.findByPhoneNumber(searchByNumberDto);
+        SearchByNumberDto searchByNumberDto = new SearchByNumberDto(phoneNumber);
+        Errors errors = new BeanPropertyBindingResult(searchByNumberDto, "searchByNumberDto");
+        searchUserByNumberValidator.validate(searchByNumberDto,errors);
+        if (errors.hasErrors()) {
+            StringBuilder sb = new StringBuilder();
+            errors.getAllErrors().forEach(err -> sb.append(messageSource.getMessage(err, Locale.ENGLISH)).append(System.lineSeparator()));
+            throw new ValidationException(sb.toString());
+        }
+        return new SearchByNumberResponseDto(userService.findByPhoneNumber(searchByNumberDto));
     }
-
 
     /**
      * Called on deleting user (user page).
@@ -100,10 +114,13 @@ public class AdminDataController {
      * @throws LogicException    if user still have contracts
      * @throws DatabaseException if user with @userId doesn't exist
      */
-    @RequestMapping(value = "/adminPanel/user/deleteUser", method = RequestMethod.POST)
-    public String deleteUser(@RequestBody String userId) throws LogicException, DatabaseException {
+    @PostMapping(value = "/adminPanel/user/deleteUser/{userId}")
+    public ResponseDto deleteUser(@PathVariable String userId) throws LogicException, DatabaseException, ValidationException {
+        if(!pathVariableIsANumber(userId)){
+            throw new ValidationException("Wrong path variable.");
+        }
         userService.deleteUserById(Integer.parseInt(userId));
-        return "User (id:" + userId + ") successfully deleted.";
+        return new ResponseDto("User (id:" + userId + ") successfully deleted.");
     }
 
     /**
@@ -114,10 +131,14 @@ public class AdminDataController {
      * @throws LogicException    if tariff is still used
      * @throws DatabaseException if tariff with @tariffId doesn't exist
      */
-    @RequestMapping(value = "/adminPanel/tariff/deleteTariff", method = RequestMethod.POST)
-    public String deleteTariff(@RequestBody @Valid String tariffId) throws DatabaseException, LogicException {
+    @PostMapping(value = "/adminPanel/tariff/deleteTariff/{tariffId}")
+    public ResponseDto deleteTariff(@PathVariable String tariffId) throws DatabaseException, LogicException, ValidationException {
+        if(!pathVariableIsANumber(tariffId)){
+            throw new ValidationException("Wrong path variable.");
+        }
         tariffService.deleteTariffById(Integer.parseInt(tariffId));
-        return "Tariff (id:" + tariffId + ") successfully deleted.";
+        return new ResponseDto("Tariff (id:" + tariffId + ") successfully deleted.");
+
     }
 
     /**
@@ -127,12 +148,14 @@ public class AdminDataController {
      * @return message
      * @throws DatabaseException if contract with @contractId doesn't exist
      */
-    @RequestMapping(value = "/adminPanel/contract/deleteContract", method = RequestMethod.POST)
-    public String deleteContract(@RequestBody String contractId) throws DatabaseException {
+    @PostMapping(value = "/adminPanel/contract/deleteContract/{contractId}")
+    public ResponseDto deleteContract(@PathVariable String contractId) throws DatabaseException, ValidationException {
+        if(!pathVariableIsANumber(contractId)){
+            throw new ValidationException("Wrong path variable.");
+        }
         contractService.deleteContractById(Integer.parseInt(contractId));
-        return "Contract (id:" + contractId + ") successfully deleted.";
+        return new ResponseDto("Contract (id:" + contractId + ") successfully deleted.");
     }
-
 
     /**
      * Deleting option.
@@ -142,10 +165,13 @@ public class AdminDataController {
      * @throws DatabaseException if option with @optionId doesn't exist
      * @throws LogicException    if options is still used somewhere.
      */
-    @RequestMapping(value = "/adminPanel/option/deleteOption", method = RequestMethod.POST)
-    public String deleteOption(@RequestBody String optionId) throws DatabaseException, LogicException {
+    @PostMapping(value = "/adminPanel/option/deleteOption/{optionId}")
+    public ResponseDto deleteOption(@PathVariable String optionId) throws DatabaseException, LogicException, ValidationException {
+        if(!pathVariableIsANumber(optionId)){
+            throw new ValidationException("Wrong path variable.");
+        }
         tariffOptionService.deleteTariffOptionById(Integer.parseInt(optionId));
-        return "Option (id:" + optionId + ") successfully deleted.";
+        return new ResponseDto("Option (id:" + optionId + ") successfully deleted.");
     }
 
     /**
@@ -158,12 +184,12 @@ public class AdminDataController {
      * @throws DatabaseException   if tariff doesn't exist
      */
     @RequestMapping(value = "/adminPanel/tariff/addOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String tariffAddOptions(@RequestBody @Valid EditTariffDto editTariffDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
+    public ResponseDto tariffAddOptions(@RequestBody @Valid EditTariffDto editTariffDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
         if (result.hasErrors()) {
             throw new ValidationException("Select options to add.");
         }
         tariffService.addOptions(editTariffDto);
-        return "Options added.";
+        return new ResponseDto("Options added.");
     }
 
     /**
@@ -175,14 +201,13 @@ public class AdminDataController {
      * @throws ValidationException if no options selected
      * @throws ValidationException if tariff doesn't exist
      */
-    @RequestMapping(value = "/adminPanel/tariff/delOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String tariffDelOptions(@RequestBody @Valid EditTariffDto editTariffDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
+    @PostMapping(value = "/adminPanel/tariff/delOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseDto tariffDelOptions(@RequestBody @Valid EditTariffDto editTariffDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
         if (result.hasErrors()) {
             throw new ValidationException("Select options to delete.");
         }
         tariffService.delOptions(editTariffDto);
-        return "Options deleted.";
-
+        return new ResponseDto("Options deleted.");
     }
 
     /**
@@ -195,12 +220,12 @@ public class AdminDataController {
      * @throws DatabaseException   if contract doesn't exist
      */
     @RequestMapping(value = "/adminPanel/contract/addOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String contractAddOptions(@RequestBody @Valid EditContractDto editContractDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
+    public ResponseDto contractAddOptions(@RequestBody @Valid EditContractDto editContractDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
         if (result.hasErrors()) {
             throw new ValidationException("Select options to add.");
         }
         contractService.adminAddOptions(editContractDto);
-        return "Options added.";
+        return new ResponseDto("Options added.");
     }
 
     /**
@@ -212,13 +237,13 @@ public class AdminDataController {
      * @throws ValidationException if no options selected
      * @throws DatabaseException   if contract doesn't exist
      */
-    @RequestMapping(value = "/adminPanel/contract/delOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String contractDelOptions(@RequestBody @Valid EditContractDto editContractDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
+    @PostMapping(value = "/adminPanel/contract/delOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseDto contractDelOptions(@RequestBody @Valid EditContractDto editContractDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
         if (result.hasErrors()) {
             throw new ValidationException("Select options to delete.");
         }
         contractService.adminDelOptions(editContractDto);
-        return "Options deleted.";
+        return new ResponseDto("Options deleted.");
     }
 
     /**
@@ -230,15 +255,14 @@ public class AdminDataController {
      * @throws ValidationException if tariff is not selected
      * @throws DatabaseException   if contract doesn't exist
      */
-    @RequestMapping(value = "/adminPanel/contract/switchTariff", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String switchTariff(@RequestBody @Valid SwitchTariffDto switchTariffDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
+    @PostMapping(value = "/adminPanel/contract/switchTariff", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE)
+    public ResponseDto switchTariff(@RequestBody @Valid SwitchTariffDto switchTariffDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
         if (result.hasErrors()) {
             throw new ValidationException("Select tariff.");
         }
         contractService.adminSwitchTariff(switchTariffDto);
-        return "Switched to tariff (id:" + switchTariffDto.getTariffId() + ").";
+        return new ResponseDto("Switched to tariff (id:" + switchTariffDto.getTariffId() + ").");
     }
-
 
     /**
      * Adding selected options to the tariff.
@@ -250,12 +274,12 @@ public class AdminDataController {
      * @throws DatabaseException   if tariff doesn't exist
      */
     @RequestMapping(value = "/adminPanel/option/addOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String optionAddOptions(@RequestBody @Valid TariffOptionRulesDto tariffOptionRulesDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
+    public ResponseDto optionAddOptions(@RequestBody @Valid TariffOptionRulesDto tariffOptionRulesDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
         if (result.hasErrors()) {
             throw new ValidationException("Select options to add.");
         }
         tariffOptionService.addRuleTariffOptions(tariffOptionRulesDto);
-        return "Options added.";
+        return new ResponseDto("Options added.");
     }
 
     /**
@@ -268,13 +292,12 @@ public class AdminDataController {
      * @throws ValidationException if tariff doesn't exist
      */
     @RequestMapping(value = "/adminPanel/option/delOptions", consumes = MediaType.APPLICATION_JSON_UTF8_VALUE, method = RequestMethod.POST)
-    public String optionDelOptions(@RequestBody @Valid TariffOptionRulesDto tariffOptionRulesDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
+    public ResponseDto optionDelOptions(@RequestBody @Valid TariffOptionRulesDto tariffOptionRulesDto, BindingResult result) throws ValidationException, DatabaseException, LogicException {
         if (result.hasErrors()) {
             throw new ValidationException("Select options to delete.");
         }
         tariffOptionService.delRuleTariffOptions(tariffOptionRulesDto);
-        return "Options deleted.";
-
+        return new ResponseDto("Options deleted.");
     }
 
     /**
@@ -285,12 +308,12 @@ public class AdminDataController {
      * @throws DatabaseException   if user doesn't exist
      */
     @RequestMapping(value = "/adminPanel/addFunds", method = RequestMethod.POST)
-    public String addFunds(@RequestBody @Valid UserFundsDto userFundsDto, BindingResult result) throws DatabaseException, ValidationException {
+    public ResponseDto addFunds(@RequestBody @Valid UserFundsDto userFundsDto, BindingResult result) throws DatabaseException, ValidationException {
         if (result.hasErrors()) {
             throw new ValidationException("Chose the amount of money you want to add (min - 10).");
         }
         userService.addFunds(userFundsDto);
-        return userFundsDto.getAmount() + " funds added.";
+        return new ResponseDto(userFundsDto.getAmount() + " funds added.");
     }
 
     /**
@@ -300,12 +323,12 @@ public class AdminDataController {
      * @throws ValidationException if amount value is null
      * @throws DatabaseException   if user doesn't exist
      */
-    @RequestMapping(value = "/adminPanel/getBalance", method = RequestMethod.POST)
-    public String getBalance(@RequestBody @Valid GetBalanceDto getBalanceDto, BindingResult result) throws DatabaseException, ValidationException {
-        if (result.hasErrors()) {
-            throw new ValidationException("Wrong input!");
+    @RequestMapping(value = "/adminPanel/getBalance/{userId}", method = RequestMethod.GET)
+    public String getBalance(@PathVariable String userId) throws DatabaseException, ValidationException {
+        if(!pathVariableIsANumber(userId)){
+            throw new ValidationException("Wrong path variable.");
         }
-        return userService.getBalance(getBalanceDto);
+        return userService.getBalance(Integer.parseInt(userId));
     }
 }
